+++ carp_mobile_sensing/lib/domain/domain.json.dart
@@ -60,8 +60,8 @@ void _registerFromJsonFunctions() {
 
   // AppTaskController classes
   FromJsonFactory().register(UserTaskSnapshotList());
-  FromJsonFactory().register(UserTaskSnapshot(
-      AppTask(type: 'ignored'), UserTaskState.canceled, DateTime.now()));
+  FromJsonFactory().register(UserTaskSnapshot(AppTask(type: 'ignored'),
+      UserTaskState.canceled, DateTime.now(), DateTime.now()));
   _fromJsonFunctionsRegistrered = true;
 }